Replace direct mention tags by people names in forwarded emails
Previously, when a Slack message included a direct mention of someone,
email sent to forward that message displayed it as a code, such as
'<@UF122JF>'. This made it impossible to know who was referred to.

This commit fixes it by replacing such tags by that user's first name,
or alternatively their full name if user's first name is unknown, or
if several users with same first name are tagged in that message.