Add ability to store a subject for a Slack thread into the database
Previously, there was no way to store the subject for a given Slack
thread. This made it impossible to ensure threading in forwarded
emails, as a reply would not get into the thread if the way to
generate a subject for a given message had changed since the first
message was sent, or if the first message had itself changed.

This commit fixes it by adding the corresponding table to the database
and creating a class to access it.