Add basic support for threading of forwarded emails
Previously, emails sent to forward Slack messages each started a new
email thread. This made it very difficult to read them, not knowing
what message was a reply to what.

This commit adds a first basic support of email threading, by adding
"Message-Id" and "In-Reply-To" email headers, as well as setting
subject to be that of first Slack message of thread.

The current limitations are:
- direct messages and group conversations are also starting a new
  thread at each message, instead of being one thread for the whole
  "channel"
- replies to messages which haven't been processed by the program, and
  thus don't appear in the database, don't include quotes of previous
  messages of the same thread
- replies to direct messages and group conversations whose first message
  has not been processed by the program, and thus haven't been entered
  into database are not supported