Add origin of message as forwarded email recipients
Previously, emails sent to forward Slack messages did not include any
information that enabled reader to assess whether message came from a
channel or not, which channel it came from, whether it was a direct
message or a group message, etc.

This made it very difficult to read and interpret, and prevented
having a user friendly convention for bot to process email replies.

This commit fixes it by specifying recipients of email as:
- only channel if message is posted to a channel
- only group members if message is posted in a group
- only recipient if message is a direct message