Add partial support for answering emails into a Slack thread
Previously, Slack messages were forwarded to user as emails. If user
answered these emails, their answer would be posted in same channel,
but not in a reply to the message, outside of any thread.

This made the service unsuited to replying by email, since threads
would be broken down at each intervention of user.

This commit fixes it partially by enabling to reply to messages by
email, sending them in the right thread in Slack.

This support is only partial because:
- as for all messages sent by email, only messages sent to a channel,
and with channel in To field, are supported
- replies to emails sent by user don't thread together, ie it is not
possible to send two emails from user in a row