Move main controller logic to its own class
Main controller logic refers to the logic that takes high level
decisions on events, such as a new Slack message received, or a
new instruction received by email, to update the system level
state, skip events or pass one to the right handler.

Previously, this logic was placed inside the main loop of the program.

This made it untestable, whereas it was needed to start making it
more complex.

This commit fixes it by moving it to a dedicated class.