Stop transferring Slack messages from main user
Previously, if a message in Slack originated from main user, it was
transferred by email like any other message. This was problematic for
two reasons.

In the short run, because it created an infinite loop, whereby forwarded
message was also received by bot, which then posted it into Slack, which
reported it to service, which transferred it to user by email, etc.

In the long run, because this meant that user could see their response
as a duplicate unread message in their email.

This commit fixes it by ignoring all Slack messages from main user.