Stop forwarding emails sent by bot to Slack
Previously, when a message was posted in Slack by a teammate of user's,
it got forwarded by email to user. But it was also forwarded to bot
email address since bot email address was included in recipients. This
mail sent by service to bot email address was then treated the same
way by the service as an email sent by user. As a consequence, service
forwarded it back to Slack, as a message posted by user.

This created a behavior where user seemed to repeat every message posted
to Slack.

This commit fixes it by ignoring emails coming from bot when processing
bot inbox.