Include tags in cc field to search for channel to post to
Previously, when an email arrived in bot inbox, service searched
for a tag in a recipient, to know which channel this email should
be forwarded to. It only looked at recipients in the To line. As
a consequence, if a channel was in a cc line, it was ignored.

This made it very cumbersome to use, and very unreliable for user,
since answering a message posted to a channel in Gmail moved the
channel address to cc.

This commit fixes it.